requirements.txt
python-binance
ConfigParser
yaml
